requirements.txt
openai>=1.0.0
PyGithub>=1.59
requests
